(refactor) Move components for editing workouts, such as editLoad and editReps into a common folder to be re-used for future doWorkout components
