+++ client/scripts/common/copyObjectHelper.js
