+++ client/scripts/common/renderTimeHelper.js
